(Osmose.java) Introduced a bug at revision 136 : deleted output folder for every replicate so was only keeping the last replicate results. Now fixed.
git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@156 45b28200-7c74-4d1a-96bc-342942c5875a

+++ Osmose/src/fr/ird/osmose/Osmose.java
@@ -254,7 +254,7 @@ public class Osmose {
                  * Delete older output directory and all its content
                  */
                 File targetPath = new File(outputPathName + outputFileNameTab[numSerie]);
-                if (targetPath.exists()) {
+                if ((numSimu == 0) && targetPath.exists()) {
                     IOTools.deleteDirectory(targetPath);
                 }
                 if (spatializedOutputs[numSerie]) {
@@ -2833,6 +2833,9 @@ public class Osmose {
         initializeOptions();
         loadMPAs();
         initializeOptions();
+        if (spatializedOutputs[numSerie]) {
+            createNCFile(numSerie);
+        }
         simulation = new Simulation();
         simulation.init();
         readMigrationFile();
